+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add github action workflow that automates publishing to npm
 - Add docs for Grid component
 - Add docs for the Modal component
+- Add docs for Heading and ControlBar components
 
 ### Changed